inline and remove serialbox_helpers.py

+++ model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/nh_solve/solve_nonhydro.py
@@ -508,7 +508,7 @@ class SolveNonhydro:
                 rd_o_p0ref=self.params.rd_o_p0ref,
                 horizontal_start=start_cell_lb,
                 horizontal_end=end_cell_nudging_minus1,
-                vertical_start=int32(0),
+                vertical_start=0,
                 vertical_end=self.grid.num_levels,
                 offset_provider={},
             )
